Update README.md
This project is a web-based scientific calculator built using Flask and developed in a Jupyter Notebook environment. It supports a wide range of mathematical operations including trigonometric functions, logarithms, exponentiation, and more. Ideal for students, educators, and developers interested in combining web development with computational tools in Python.

With Flask serving as the backend and interactive widgets for the frontend, this project showcases how lightweight web applications can be created and tested directly within Jupyte